requirements.txt
alembic==1.7.5
aniso8601==9.0.1
astroid==2.9.3
attrs==21.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coverage==6.3
coveralls==3.3.1
cryptography==36.0.1
docopt==0.6.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.29
threadpoolctl==3.0.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
